refactor: add iroh-base crate
