+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
   "iroh",
   "iroh-bytes",
+  "iroh-base",
   "iroh-gossip",
   "iroh-metrics",
   "iroh-net",
