+++ iroh-base/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "iroh-base"
+version = "0.10.0"
+edition = "2021"
+readme = "README.md"
+description = "base type and utilities for Iroh"
+license = "MIT OR Apache-2.0"
+authors = ["n0 team"]
+repository = "https://github.com/n0-computer/iroh"
+
+# Sadly this also needs to be updated in .github/workflows/ci.yml
+rust-version = "1.72"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { version = "1", features = ["backtrace"] }
+bao-tree = { version = "0.9.1", features = ["tokio_fsm"], default-features = false }
+bytes = { version = "1.4", features = ["serde"] }
+data-encoding = "2.3.3"
+hex = "0.4.3"
+multibase = "0.9.1"
+postcard = { version = "1", default-features = false, features = ["alloc", "use-std", "experimental-derive"] }
+serde = { version = "1", features = ["derive"] }
+serde-error = "0.1.2"
+thiserror = "1"
+
+[dev-dependencies]
+
+[features]
+default = ["hash", "base32"]
+hash = []
+base32 = []
