+++ iroh-base/src/lib.rs
@@ -1,0 +1,7 @@
+//! Base types and utilities for Iroh
+
+#[cfg(feature = "base32")]
+pub mod base32;
+#[cfg(feature = "hash")]
+pub mod hash;
+pub mod rpc;
