+++ iroh-bytes/Cargo.toml
@@ -25,8 +25,8 @@ flume = "0.11"
 futures = "0.3.25"
 genawaiter = { version = "0.99.1", features = ["futures03"] }
 hex = "0.4.3"
+iroh-base = { version = "0.10.0", path = "../iroh-base" }
 iroh-io = { version = "0.3.0", features = ["stats"] }
-multibase = "0.9.1"
 num_cpus = "1.15.0"
 once_cell = "1.17.0"
 postcard = { version = "1", default-features = false, features = ["alloc", "use-std", "experimental-derive"] }
