+++ iroh-bytes/src/store/traits.rs
@@ -5,14 +5,15 @@ use crate::{
     hashseq::parse_hash_seq,
     util::{
         progress::{IdGenerator, ProgressSender},
-        BlobFormat, HashAndFormat, RpcError, Tag,
+        Tag,
     },
-    Hash, TempTag,
+    BlobFormat, Hash, HashAndFormat, TempTag,
 };
 use bao_tree::{blake3, ChunkRanges};
 use bytes::Bytes;
 use futures::{future::BoxFuture, stream::LocalBoxStream, Stream, StreamExt};
 use genawaiter::rc::{Co, Gen};
+use iroh_base::rpc::RpcError;
 use iroh_io::AsyncSliceReader;
 use serde::{Deserialize, Serialize};
 use tokio::{io::AsyncRead, sync::mpsc};
