+++ iroh-net/Cargo.toml
@@ -29,6 +29,7 @@ ed25519-dalek = { version = "2.0.0", features = ["serde", "rand_core"] }
 flume = "0.11"
 futures = "0.3.25"
 governor = "0.6.0"
+iroh-base = { version = "0.10.0", path = "../iroh-base" }
 hex = "0.4.3"
 hostname = "0.3.1"
 http = "0.2.9"
