+++ iroh-sync/Cargo.toml
@@ -22,6 +22,7 @@ derive_more = { version = "1.0.0-beta.1", features = ["debug", "deref", "display
 ed25519-dalek = { version = "2.0.0", features = ["serde", "rand_core"] }
 flume = "0.11"
 iroh-bytes = { version = "0.10.0", path = "../iroh-bytes" }
+iroh-base = { version = "0.10.0", path = "../iroh-base" }
 iroh-metrics = { version = "0.10.0", path = "../iroh-metrics", optional = true }
 num_enum = "0.7"
 once_cell = "1.18.0"
