+++ iroh/Cargo.toml
@@ -26,6 +26,7 @@ futures = "0.3.25"
 genawaiter = { version = "0.99", default-features = false, features = ["futures03"] }
 hex = { version = "0.4.3" }
 iroh-bytes = { version = "0.10.0", path = "../iroh-bytes" }
+iroh-base = { version = "0.10.0", path = "../iroh-base" }
 iroh-io = { version = "0.3.0", features = ["stats"] }
 iroh-metrics = { version = "0.10.0", path = "../iroh-metrics", optional = true }
 iroh-net = { version = "0.10.0", path = "../iroh-net" }
