+++ iroh/src/lib.rs
@@ -1,6 +1,7 @@
 //! Send data over the internet.
 #![deny(missing_docs, rustdoc::broken_intra_doc_links)]
 
+pub use iroh_base as base;
 pub use iroh_bytes as bytes;
 pub use iroh_net as net;
 pub use iroh_sync as sync;
